# Add caching to reduce runtime by 85% (all files) and 42% (single file)