@AlexEvanczuk 2021-12-20
I did some more first principles analysis of performance with packwerk and wanted to share some of the results.

I used a flamegraph to record `bin/packwerk check` (in serial so it could more accurately sample the work in the same process).

Here's what I ran to get this flamegraph:
```
# Get rbspy, the tool used to sample a ruby process
wget https://github.com/rbspy/rbspy/releases/download/v0.9.1/rbspy-x86_64-unknown-linux-gnu.tar.gz
# Unpack the file
tar xvf rbspy-x86_64-unknown-linux-gnu.tar.gz
cp rbspy-x86_64-unknown-linux-gnu /usr/local/bin/rbspy
chmod +x /usr/local/bin/rbspy
# Run without spring, as I believe spring pulls some things into a separate process which affects our ability to accurately sample
DISABLE_SPRING=1 /usr/local/bin/rbspy record --format speedscope --file data.json -- bin/packwerk check
# Tool used to view the flamegraph
npm install -g speedscope
speedscope data.json
```

## Results in time order
<img width="1680" alt="Screen Shot 2021-12-20 at 12 10 19 PM" src="https://user-images.githubusercontent.com/3311200/146805930-7817f53e-dc0b-4055-a6ee-1eec7c45bd18.png">

## Results in left heavy
<img width="1680" alt="Screen Shot 2021-12-20 at 12 12 50 PM" src="https://user-images.githubusercontent.com/3311200/146805996-09598af1-f55f-4321-ad9f-4f4e7b6f41a1.png">

## Analysis
In general, we can see here that the vast majority of the time is spent in two steps:
1) Parsing the ruby file
2) Traversing through the tree and collecting references.

Note that today, collecting references uses the constant resolver, and we create a custom packwerk 'reference' to establish an edge. Note that if we were to ever use a sorbet parser, we would need to take in a parsed AST that doesn't know about references/offenses and turn it into something else. So I think this could be a good refactor that doubles as giving us something that is much simpler to cache.