@AlexEvanczuk 2021-12-21
I tried this other implementation here: https://github.com/Shopify/packwerk/pull/168

The result I believe is much simpler.